added algebra project
