+++ algebra/src/lib.rs
@@ -1,0 +1,3 @@
+mod algebraic_extensions;
+
+pub use algebraic_extensions::{NumericField, NumericRing};
